# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Raj's Austin Events newsletter
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /newsletter/subscribe:
    post:
      operationId: subscribeToNewsletter
      tags: [newsletter]
      summary: Subscribe to the newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: Successfully subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /newsletter/unsubscribe:
    post:
      operationId: unsubscribeFromNewsletter
      tags: [newsletter]
      summary: Unsubscribe from the newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
      responses:
        "200":
          description: Successfully unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /tenant/config:
    get:
      operationId: getTenantConfig
      tags: [tenants]
      summary: Get public config for a tenant by slug
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
          description: The tenant's subdomain slug (e.g. "austin")
      responses:
        "200":
          description: Tenant config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantConfigResponse"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
      required:
        - email
    SubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        subscriber:
          $ref: "#/components/schemas/Subscriber"
      required:
        - success
        - message
    UnsubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    Subscriber:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        subscribedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - email
        - subscribedAt
        - isActive
...
    TenantPublic:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        city:
          type: string
        accentColor:
          type: string
        categories:
          type: array
          items:
            type: string
      required:
        - slug
        - name
        - city
        - accentColor
        - categories
    TenantConfigResponse:
      type: object
      properties:
        tenant:
          $ref: "#/components/schemas/TenantPublic"
      required:
        - tenant
...